Massive commit touching a bit of everything on the verge of the 1.6 release. See the different discussions on gama-dev and issues fixed since the latest release.
git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@6282 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a

+++ msi.gama.application/src/msi/gama/gui/parameters/EditorFactory.java
@@ -119,7 +119,7 @@ public class EditorFactory implements IEditorFactory {
 	public AbstractEditor create(final IAgent agent, final IParameter var, final EditorListener l) {
 		final boolean canBeNull =
 			var instanceof ExperimentParameter ? ((ExperimentParameter) var).canBeNull() : false;
-		final short type = var.getType().id();
+		final int type = var.getType().id();
 		AbstractEditor gp =
 			var.getType().isSpeciesType() | type == IType.AGENT ? new AgentEditor(agent, var, l)
 				: type == IType.BOOL ? new BooleanEditor(agent, var, l) : type == IType.COLOR
